% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioStatR-package.R
\docType{package}
\name{BioStatR-package}
\alias{BioStatR-package}
\alias{BioStatR}
\title{BioStatR: Initiation à La Statistique Avec R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Datasets and functions for the book "Initiation à la Statistique avec R", F. Bertrand and M. Maumy-Bertrand (2022, ISBN:978-2100782826 Dunod, fourth edition).
}
\references{
F. Bertrand, M. Maumy-Bertrand, Initiation à la Statistique avec
R, 4ème édition, ISBN:9782100847945, Dunod, Paris, 2023

\emph{Initiation à la Statistique avec R}, Frédéric Bertrand, 
Myriam Maumy-Bertrand, 2023, ,
\url{https://www.dunod.com/sciences-techniques/initiation-statistique-avec-r-cours-exemples-exercices-et-problemes-corriges-1},
\url{https://github.com/fbertran/BioStatR/} et
\url{https://fbertran.github.io/BioStatR/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fbertran.github.io/BioStatR/}
  \item \url{https://github.com/fbertran/BioStatR/}
  \item Report bugs at \url{https://github.com/fbertran/BioStatR/issues/}
}

}
\author{
\strong{Maintainer}: Frederic Bertrand \email{frederic.bertrand@lecnam.net} (\href{https://orcid.org/0000-0002-0837-8281}{ORCID})

Authors:
\itemize{
  \item Myriam Maumy-Bertrand \email{myriam.maumy@ehesp.fr} (\href{https://orcid.org/0000-0002-4615-1512}{ORCID})
}

}
\keyword{internal}
