% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioStatR-package.R
\docType{package}
\name{BioStatR}
\alias{BioStatR}
\title{BioStatR}
\description{
Motivation: Package compagnon du livre Initiation à la statistique avec R.
Il contient les codes des chapitres du livre ainsi que les solutions des
exercices mais aussi d'autres compléments à découvrir.
}
\examples{
set.seed(314)

}
\references{
F. Bertrand, M. Maumy-Bertrand, Initiation à la Statistique avec
R, 4ème édition, ISBN:9782100847945, Dunod, Paris, 2023

\emph{Initiation à la Statistique avec R}, Frédéric Bertrand, 
Myriam Maumy-Bertrand, 2023, ,
\url{https://www.dunod.com/sciences-techniques/initiation-statistique-avec-r-cours-exemples-exercices-et-problemes-corriges-1},
\url{https://github.com/fbertran/BioStatR/} et
\url{https://fbertran.github.io/BioStatR/}
}
