% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvar.R
\name{cvar}
\alias{cvar}
\title{Coefficient de variation}
\usage{
cvar(x)
}
\arguments{
\item{x}{Un vecteur numérique}
}
\value{
\item{num}{Valeur du coefficient de variation exprimé en pourcents}
}
\description{
Calcule coefficent de variation d'une série statistique
}
\details{
Le coefficient de variation est égal à l'écart-type corrigé divisé par la
moyenne. Il est exprimé en pourcents.
}
\examples{

data(Europe)
cvar(Europe[,2])

}
\references{
F. Bertrand, M. Maumy-Bertrand, Initiation à la Statistique avec
R, Dunod, 3ème edition, 2018.
}
\seealso{
\code{\link{mean}}, \code{\link{sd}}
}
\author{
Frédéric Bertrand\cr \email{frederic.bertrand@utt.fr}\cr
\url{http://www-irma.u-strasbg.fr/~fbertran/}\cr 
Maumy-Bertrand\cr \email{myriam.maumy@utt.fr}\cr
\url{http://www-irma.u-strasbg.fr/~mmaumy/}
}
\keyword{univar}
