% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta2.R
\name{eta2}
\alias{eta2}
\title{Calcul du rapport de corrélation eta carré}
\usage{
eta2(x, y)
}
\arguments{
\item{x}{Un vecteur associé à la variable quantitative}

\item{y}{Un facteur associé à la variable qualitative}
}
\value{
\item{num}{La valeur du rapport de corrélation empirique}
}
\description{
Cette fonction calcule le rapport de corrélation \eqn{\eta^2} qui est une
mesure d'association importante entre une variable quantitative et une
variable qualitative.
}
\examples{

eta2(Mesures5$taille,Mesures5$espece)

}
\references{
F. Bertrand, M. Maumy-Bertrand, Initiation à la Statistique avec
R, Dunod, 4ème édition, 2023.
}
\author{
Frédéric Bertrand\cr \email{frederic.bertrand@utt.fr}\cr
\url{http://www-irma.u-strasbg.fr/~fbertran/}\cr 
Maumy-Bertrand\cr \email{myriam.maumy@utt.fr}\cr
\url{http://www-irma.u-strasbg.fr/~mmaumy/}
}
\keyword{univar}
