\name{Extrait_Taille}
\alias{Extrait_Taille}
\docType{data}
\title{Mesures de fruits d'arbustes}
\description{
This dataset measurements of several features of the fruits of small trees such as their mass (in g) or their length (in cm).
}
\usage{Extrait_Taille}
\format{
  A data frame with 252 observations on the following 5 variables.
  \describe{
    \item{\code{masse}}{a numeric vector}
    \item{\code{taille}}{a numeric vector}
    \item{\code{espece}}{a factor with levels \code{bignone}, \code{glycine blanche}, \code{glycine violette} and \code{lauriers roses}}
  }
}
\details{
This dataset was made during the summer 2009 in the south of France. It provides measurements of several features of the fruits of small trees such as their mass or their length.
}
\source{F. Bertrand, M. Maumy-Bertrand, Initiation à la Statistique avec R, Dunod, 2010.}
\examples{
data(Extrait_Taille)
}
\keyword{datasets}
