% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Obtencion de la tabla de frecuencias absolutas y relativas de las categorias de una variable}
\usage{
freq(x = NULL, acum = TRUE, cuts = 0, agrup = TRUE, decs = 3, grf = TRUE, ...)
}
\arguments{
\item{x}{vector: vector o data.frame a describir}

\item{acum}{valor logico: TRUE=proporciona la frecuencia relativa acumulada}

\item{cuts}{valor entero: numero de intervalos a realizar. Si se omite se utiliza el criterio de Sturges}

\item{agrup}{valor logico: si agrup=FALSE no se hace agrupacion en intervalos aunque haya mas de 10 categorias}

\item{decs}{valor entero: numero de decimales a mostrar en la salida}

\item{grf}{valor logico: si TRUE/FALSE se proporciona/omite salida grafica}

\item{...}{parametros de configuracion de la funcion grpsggp}
}
\value{
si x es un vector, se devuelve un data.frame con la tabla de frecuencias. Si x es un data.frame, se muestra la tabla de frecuencias de cada columna, pero la funcion no devuelve ningun objeto
}
\description{
Permite obtener la tabla de frecuencias observadas para un vector de datos x o cada columna de un data.frame x (Texto intencionadamente sin tildes u otros caracteres especiales por la incompatibilidad de los mapas de caracteres).
}
\examples{
dat<-c(12,15,13,12,11,14,15,15,15,12,11,13,14,15,NA)
freq(dat)

cats<-c('a','b','c','b','c','b','c','a','c','c','a','a','a')
freq(cats,acum=FALSE,grf=TRUE)

dat2<-rnorm(550,212.3,6.3)
freq(dat2, agrup=TRUE,cuts=5)

t<-rbinom(25,20,0.65)
freq(t,agrup=FALSE,cuts=5,decs=2)

nrm<-rnorm(50,250,2)
bnm<-rbinom(50,80,0.5)
df<-data.frame(nrm,bnm)
freq(nrm)
freq(bnm,agrup=TRUE,grf=TRUE)
freq(df,acum=FALSE,grf=TRUE,hnmin=60)


}
