% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_utils.R
\name{get_msg}
\alias{get_msg}
\title{Get multilingual message}
\usage{
get_msg(key)
}
\arguments{
\item{key}{Message key}
}
\value{
Message string
}
\description{
Returns a message in the language specified by option 'BioStatR.lang' (default 'es').
}
