% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpsggp.R
\name{grpsggp}
\alias{grpsggp}
\title{Diagramas por grupos}
\usage{
grpsggp(
  x = NULL,
  f = NULL,
  se = NULL,
  ggid = 1,
  lbls = NULL,
  bins = NULL,
  hnmin = 50
)
}
\arguments{
\item{x}{vector: variable cuantitativa a resumir}

\item{f}{vector: factor cuyos niveles segmentan a la variable cuantitativa (si falta se da la descriptiva de x)}

\item{se}{valor real: error estandar a representar}

\item{ggid}{valor entero: indicador del grafico a representar}

\item{lbls}{vector de caracteres: nombres de los ejees}

\item{bins}{entero: numero de intervalos para el histograma}

\item{hnmin}{entero: tamano minimo/maximo de muestra para representar el histograma/diagrama de puntos}
}
\value{
Representacion grafica
}
\description{
Obtencion de los diagramas de distribucion segmentados, o no, por factor de agrupacion
}
