% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icl.R
\name{icl}
\alias{icl}
\title{Estimacion del parametro de la distribucion de Poisson}
\usage{
icl(x = 0, n = 0, conf = 0.95, alfa = 1 - conf, decs = 4, d = 0, eco = TRUE)
}
\arguments{
\item{x}{valor entero: vector de observaciones de una v.a. con distribucion de Poisson o valor medio observado en una muestra de tamano n.}

\item{n}{valor entero: tamano de la muestra piloto cuando x representa su media.}

\item{conf}{valor < 1: nivel de confianza (alternativo al error alfa, en tanto por uno). Por defecto =.95.}

\item{alfa}{valor < 1: error alfa (alternativo al nivel de confianza, en tanto por uno). Por defecto =.05.}

\item{decs}{valor entero: precision decimal para la salida de resultados. Por defecto = 4.}

\item{d}{valor real < 1: precision deseada para el intervalo de confianza. Si d>0 se invoca a la funcion nl()}

\item{eco}{valor logico: si eco=TRUE la funcion genera un informe (no devuelve valores), si eco=FALSE devuelve el IC exacto y el basado en la aprox normal con su precision respectiva}
}
\value{
Informe con intervalos de confianza exacto y aproximado a la normal (con cpc) (eco=TRUE) o los limites y precision correspondientes a cada intervalo
}
\description{
Obtencion del intervalo de confianza para el parametro de la distribucion de Poisson por los metodos exacto y aproximando a la normal (transformacion de la raiz).
}
\examples{
# Introduciendo datos observados
# Una sola observacion (muestra de tamano n=1)
icl(3)
# muestra con mas de una observacion
icl(c(3,6,3,1,2,5))
# Introduciendo media x de n datos observados
icl(x=25, n=210)

icl(x=25, n=210, eco=FALSE)->IC
IC

# solicitud del tamano de muestra necesario para estimar con una precision d=1 unidad
icl(x=25, n=210, d=1)

}
