% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.R
\name{icm}
\alias{icm}
\title{Estimacion de la media de una variable aleatoria con distribucion normal}
\usage{
icm(
  x = 0,
  n = 0,
  m = 0,
  s = 0,
  conf = 0.95,
  alfa = 0.05,
  decs = 3,
  d = 0,
  vac = TRUE,
  eco = TRUE
)
}
\arguments{
\item{x}{vector de datos cuya media se va a estimar}

\item{n}{valor entero: tamano muestral cuando se indican los datos resumidos}

\item{m}{valor real: media de la variable (previamente calculada)}

\item{s}{valor real: desviacion tipica de la variable  (previamente calculada)}

\item{conf}{valor < 1: nivel de confianza (parametro alternativo al error alfa, en tanto por uno). Por defecto =.95.}

\item{alfa}{valor < 1: error alfa (parametro alternativo al nivel de confianza, en tanto por uno). Por defecto =.05.}

\item{decs}{valor entero: precision decimal para la salida de resultados. Por defecto = 4.}

\item{d}{valor real < 1: precision deseada para el intervalo de confianza. Si d>0 se invoca a la funcion nm() para estimar el tamano de muestra}

\item{vac}{valor logico: TRUE=se trata de una variable aleatoria continua; FALSE= la variable es discreta y se aplica cpc. Por defecto = TRUE.}

\item{eco}{valor logico: si eco=TRUE la funcion genera un informe (no devuelve valores), si eco=FALSE devuelve los limites del IC y su precision}
}
\value{
Informe (si eco=T) con el intervalo de confianza para estimar la media poblacional de una variable aleatoria normal. Limites inferior y superior del IC y su precision.
}
\description{
Permite obtener el intervalo de confianza a partir de una variable o bien de las medidas resumidas. Texto intencionadamente sin tildes u otros caracteres especiales por la incompatibilidad de los mapas de caracteres.
}
\examples{
icm(x=c(25.4, 14.6, 23.1, 26.0, 14.4, 24.3, 36.1, 21.0, NA, 41.9))
icm(x=c(25,14,23,26,14,24,36,21,NA,41), vac=FALSE)
icm(n=100, m=25.3, s=4.1)
icm(n=100, m=25.3, s=4.1, conf=0.99)
icm(n=100, m=25.3, s=4.1, alfa=0.01)

icm(n=100, m=25.3, s=4.1, alfa=.01, eco=FALSE)->IC
IC

}
