% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icp.R
\name{icp}
\alias{icp}
\title{Estimacion de una proporcion binomial por los metodos: Wilson (con cpc), Wald (con cpc) y Wald ajustado}
\usage{
icp(
  x = 0,
  n = 0,
  level = "",
  conf = 0.95,
  alfa = 0.05,
  decs = 4,
  d = 0,
  tabla = FALSE
)
}
\arguments{
\item{x}{valor entero: numerador de la proporcion binomial (si se da la informacion resumida) o valor a seleccionar del vector v (si se ha dado un vector de datos)}

\item{n}{valor entero: denominador de la proporcion binomial (si se da la informacion resumida)}

\item{level}{entero o texto: es el valor o la etiqueta del nivel de seleccion de x cuando este ultimo es un vector}

\item{conf}{valor < 1: nivel de confianza (alternativo al error alfa, en tanto por uno). Por defecto =.95.}

\item{alfa}{valor < 1: error alfa (alternativo al nivel de confianza, en tanto por uno). Por defecto =.05.}

\item{decs}{valor entero: precision decimal para la salida de resultados. Por defecto = 4.}

\item{d}{valor real < 1: precision deseada para el intervalo de confianza. Si d>0 se invoca a la funcion np()}

\item{tabla}{valor logico: si tabla=TRUE el informe tiene forma de tabla. En cualquier caso devuelve los limites del IC y su precision}
}
\value{
Informe con los intervalos de confianza de Wilson, Wald (ambos con cpc) y Wald ajustado. Limites de cada intervalo y su precision en forma de tabla.
}
\description{
Obtencion del intervalo de confianza para una proporcion binomial considerando los metodos de Wilson (con cpc), Wald (con cpc) y Wald ajustado. Texto intencionadamente sin tildes u otros caracteres especiales por la incompatibilidad de los mapas de caracteres
}
\examples{
# Introduciendo frecuencias
icp(x=25, n=210)
icp(x=25, n=210, conf=0.90, decs=8)

# Introduciendo datos
datos<-c(1,1,1,2,2,2,2,2,1,1,1,2,1,2,2,2,2,2,1,1,1,2,1,2,1,2,1,2,2,2,2,2,1,1)
  icp(x=datos, level=1)
  icp(x=datos, level=1, conf=.99)
sexo<-as.factor(c("H","H","M","M","H","M","M","H","H","M","H","M","H"))
  icp(x=sexo, level="M")

# Cambiar la salida por una tabla resumen (manejable como data.frame)
icp(x=sexo, level="M",tabla=TRUE)

# Invocacion a la estimacion del tamano muestral
icp(x=sexo, level="M", d=0.1)

}
