% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Mesures}
\alias{Mesures}
\title{Mesures de fruits d'arbustes}
\format{
A data frame with 252 observations on the following 3 variables.
\describe{ \item{masse}{a numeric vector} \item{taille}{a
numeric vector} \item{espece}{a factor with levels \code{bignone},
\code{glycine blanche}, \code{glycine violette} and \code{lauriers roses}} }
}
\description{
This dataset measurements of several features of the fruits of small trees
such as their mass or their length.
}
\details{
This dataset was made during the summer 2009 in the south of France. It
provides measurements of several features of the fruits of small trees such
as their mass or their length.
}
\examples{

data(Mesures)

}
\references{
F. Bertrand, M. Maumy-Bertrand, Initiation à la Statistique avec
R, Dunod, 4ème édition, 2023.
}
\keyword{datasets}
