% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl.R
\name{nl}
\alias{nl}
\title{Tamano de muestra para estimar el parametro de la distribucion de Poisson}
\usage{
nl(
  x = NA,
  n = 0,
  d = 0,
  lmax = 0,
  conf = 0.95,
  alfa = 0.05,
  decs = 3,
  eco = TRUE
)
}
\arguments{
\item{x}{valor entero: observacion o vector de observaciones de una v.a. con distribucion de Poisson o valor medio observado en una muestra de tamano n}

\item{n}{valor entero: tamano de la muestra piloto cuando x representa su media}

\item{d}{valor real entero: precision deseada para estimar el parametro de la distribucion de Poisson}

\item{lmax}{valor real: maximo valor del parametro de la Poisson cuando se cuenta con esta informacion}

\item{conf}{valor < 1: nivel de confianza (alternativo al error alfa, en tanto por uno). Por defecto =.95.}

\item{alfa}{valor < 1: error alfa (alternativo al nivel de confianza, en tanto por uno). Por defecto =.05.}

\item{decs}{valor entero: precision decimal para la salida de resultados.}

\item{eco}{valor logico: si eco=TRUE devuelve informe, de lo contrario valores estimados de n, sin y con cpc}
}
\value{
Estimacion del tamano de muestra sin y con correccion por continuidad (cpc)
}
\description{
Estimacion del tamano de muestra para estimar el parametro de la distribucion de Poisson con la precision deseada
}
\examples{
# Introduciendo datos observados
# una observacion (muestra de tamano 1). Precision deseada de 1 unidad
nl(3, d=1)
# si se dispone del valor maximo del parametro (precision = 1 unidad)
nl(lmax=4.5, d=1)
# muestra con mas de una observacion
nl(c(3,6,3,1,2,5),d=1)
# Introduciendo media x de n datos observados
nl(x=25, n=210, d=2)


}
