% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{nm}
\alias{nm}
\title{Tamano de muestra para estimar la media de una variable aleatoria con distribucion normal con una precision determinada}
\usage{
nm(
  x = 0,
  n = 0,
  m = 0,
  s = 0,
  d = 0,
  conf = 0.95,
  alfa = 0.05,
  decs = 4,
  vac = TRUE,
  eco = TRUE
)
}
\arguments{
\item{x}{vector de datos de la muestra piloto}

\item{n}{valor entero: tamano muestral de la muestra piloto cuando se indican sus parametros muestrales resumidos}

\item{m}{valor real: media de la de la muestra piloto (previamente calculada)}

\item{s}{valor real: desviacion tipica de la muestra piloto  (previamente calculada)}

\item{d}{valor real: precision deseada para el intervalo de confianza}

\item{conf}{valor real < 1: nivel de confianza (parametro alternativo al error alfa, en tanto por uno). Por defecto =.95.}

\item{alfa}{valor real < 1: error alfa (parametro alternativo al nivel de confianza, en tanto por uno). Por defecto =.05.}

\item{decs}{valor entero: precision decimal para la salida de resultados. Por defecto = 4.}

\item{vac}{valor logico: TRUE=se trata de una variable aleatoria continua; FALSE= la variable es discreta y se aplica cpc. Por defecto = TRUE.}

\item{eco}{valor logico: si eco=TRUE la funcion genera un informe (no devuelve valores), si eco=FALSE devuelve el tama?o de muestra estimado}
}
\value{
Informe (si eco=T) con el tamano de muestra estimado para obtener un intervalo de confianza para estimar una media con la precision deseada
}
\description{
Permite estimar el tamano muestral para estimar una media con una precision deseada a partir de la informacion dada por una muestra piloto, que puede aparecer como variable o bien dando sus medidas resumidas. Texto intencionadamente sin tildes u otros caracteres especiales por la incompatibilidad de los mapas de caracteres.
}
\examples{
nm(x=c(25.4, 14.6, 23.1, 26.0, 14.4, 24.3, 36.1, 21.0, NA, 41.9,NA), d=1.5)
nm(d=1.0, n=100, m=25.3, s=4.1)
nm(d=1.0, n=100, m=25.3, s=4.1, conf=.99)
nm(d=1.0, n=100, m=25.3, s=4.1, alfa=.01)

}
