% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np.R
\name{np}
\alias{np}
\title{Tamano de muestra para estimar una proporcion binomial}
\usage{
np(
  x = 0,
  n = 0,
  d = 0,
  v = 0,
  level = "",
  conf = 0.95,
  alfa = 0.05,
  decs = 4,
  eco = TRUE
)
}
\arguments{
\item{x}{valor entero: numerador de la proporcion binomial (si se da la informacion resumida) o valor de seleccion (si se da un vector)}

\item{n}{valor entero: denominador de la proporcion binomial (si se da la informacion resumida)}

\item{d}{valor real < 1: precision deseada para el intervalo de confianza}

\item{v}{vector: numero de casos favorables (numerador de la proporcion binomial)}

\item{level}{texto: si se indica v como factor, level es la etiqueta del nivel de seleccisn}

\item{conf}{valor < 1: nivel de confianza (parametro alternativo al error alfa, en tanto por uno). Por defecto =.95.}

\item{alfa}{valor < 1: error alfa (parametro alternativo al nivel de confianza, en tanto por uno). Por defecto =.05.}

\item{decs}{valor entero: precision decimal para la salida de resultados. Por defecto = 4.}

\item{eco}{valor logico: si eco=TRUE devuelve informe, de lo contrario, los valores estimados de n con y sin informacion}
}
\value{
Informe con el tamano de muestra necesario para estimar un proporcion binomial con la precision deseada
}
\description{
Obtencion del tamano muestral para estimar una proporcion binomial con la precision deseada a partir de informacion piloto (metodo de Wald ajustado) o sin ella. Texto intencionadamente sin tildes u otros caracteres especiales por la incompatibilidad de los mapas de caracteres.
}
\examples{
np(x=25, n=210, d=.05,decs=3)
np(x=115, n=210, d=.10)
np(x=25, n=210, d=.05, conf=.90, decs=5)
np(x=25, n=210, d=.05, conf=.90, decs=5, eco=FALSE)

}
