% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Quetelet}
\alias{Quetelet}
\title{Indices de Quetelet}
\format{
Un data frame avec 66 observations de 3 variables.  \describe{
\item{sexe}{un facteur donnant le sexe de l'individu}
\item{poids}{le poids de l'individu} \item{taille}{la
hauteur de l'individu} }
}
\description{
Ce jeu de données contient des mesures de masse et de taille pour permettre
le calcul de l'indice de masse corporelle (aussi dit de Quetelet).
}
\details{
Le poids est exprimée en kg et la hauteur en cm
}
\examples{

data(Quetelet)

}
\references{
F. Bertrand, M. Maumy-Bertrand, Initiation à la Statistique avec
R, Dunod, 4ème édition, 2023.
}
\keyword{datasets}
