% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlogitm.R
\name{rlogitm}
\alias{rlogitm}
\title{Regresion logistica multivariable}
\usage{
rlogitm(
  f,
  data = NULL,
  pred = NULL,
  grf = FALSE,
  alfa = 0.05,
  conf = 1 - alfa,
  decs = 3
)
}
\arguments{
\item{f}{formula: especificacion del modelo (ej. y ~ x1 + x2)}

\item{data}{data.frame: tabla de datos}

\item{pred}{data.frame: valores de los regresores para realizar pronosticos a partir del modelo}

\item{grf}{valor logico: si grf=TRUE se incluye el grafico de la curva ROC. Por defecto = FALSE.}

\item{alfa}{valor real < 1: error alfa (parametro alternativamente al nivel de confianza, en tanto por uno). Por defecto =.05.}

\item{conf}{valor real < 1: nivel de confianza para la elaboracion del IC para la estimacion del efecto. Por defecto = 1-alfa.}

\item{decs}{valor entero: precision decimal para la salida de resultados. Por defecto = 3.}
}
\value{
Informe con medidas de asociacion (OR), estimacion de los parametros de regr. logistica multivariable, bondad de ajuste (Hosmer-Lemeshow) y curva ROC.
}
\description{
Ajuste del modelo de regresion logistica multivariable de acuerdo a la especificacion y ~ x1 + x2 + ...
}
\examples{
# Ejemplo 1 - Uso basico
data(osteo)
rlogitm(osteo_cue ~ imc + edad, data = osteo)

# Ejemplo 2 - Con predicciones
data(osteo)
nuevos_datos <- data.frame(imc = c(22, 25), edad = c(30, 50))
rlogitm(osteo_cue ~ imc + edad, data = osteo, pred = nuevos_datos)
}
