% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rls.R
\name{rls}
\alias{rls}
\title{Regresion lineal simple}
\usage{
rls(
  y = NULL,
  x = NULL,
  data = NULL,
  pred = NULL,
  grf = TRUE,
  dfout = F,
  alfa = 0.05,
  conf = 1 - alfa,
  decs = 3
)
}
\arguments{
\item{y}{vector: variable explicada o formula del modelo lineal}

\item{x}{vector: variable explicativa}

\item{data}{data.frame: tabla de datos (necesaria si se indica y como formula)}

\item{pred}{vector: valores del regresor para realizar pronosticos a partir del modelo}

\item{grf}{valor logico: si grf=FALSE se omite la salida grafica}

\item{dfout}{valor logico: si dfout=TRUE el procedimiento devuelve la matriz de datos con valores residuales y pronosticos}

\item{alfa}{valor real < 1: error alfa (parametro alternativamente al nivel de confianza, en tanto por uno). Por defecto =.05.}

\item{conf}{valor real < 1: nivel de confianza para la elaboracion del IC para la estimacion del efecto}

\item{decs}{valor entero: precision decimal para la salida de resultados. Por defecto = 4.}
}
\value{
Informe con medidas descriptivas, correlaciones, estimacion de los parametros de regr. lineal, descriptiva residual y graficos de dispersion y de la distribucion residual
}
\description{
Ajuste del modelo de regresion lineal simple de acuerdo a la especificacion y~x
}
\examples{
pre  <-c(200.1, 190.9, 192.7, 213, 241.4, 196.9, 172.2, 185.5, 205.2, 193.7)
post <-c(392.9, 393.2, 345.1, 393, 434, 427.9, 422, 383.9, 392.3, 352.2)
# Ejemplo 1 - entrada como vectores individuales
rls(post, pre) #ajusta el modelo post~pre
# Ejemplo 2 - entrada como formula
datos<-data.frame(pre,post)
rls(post~pre, data=datos)
# Ejemplo 3 - Pronosticos
rls(post~pre, data=datos,pred=c(197,205))
# Ejemplo 4 - Obtencion de la matriz con residuos y pronosticos
tabla<-rls(post~pre, data=datos,dfout=TRUE)
head(tabla)
}
