% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2x2.R
\name{tabla2x2}
\alias{tabla2x2}
\title{Analisis de una tabla de contingencia 2x2}
\usage{
tabla2x2(
  frecs = NULL,
  fvar = NULL,
  cvar = NULL,
  o = NULL,
  o11 = 0,
  o12 = 0,
  o21 = 0,
  o22 = 0,
  fcat = c("F1", "F2"),
  ccat = c("C1", "C2"),
  estudio = "T",
  tablas = "",
  medidas = TRUE,
  alfa = 0.05,
  conf = 0.95,
  decs = 3,
  ptol = 0.001
)
}
\arguments{
\item{frecs}{data.frame: data.frame con frecuencias observadas}

\item{fvar}{vector de enteros o factores: Variable por filas}

\item{cvar}{vector de enteros o factores: Variable por columnas}

\item{o}{vector de enteros: Vector de frecuencias observadas con la organizacion (o11,o12,o21,o22)}

\item{o11}{entero: frecuencia observada en la fila 1 y columna 1}

\item{o12}{entero: frecuencia observada en la fila 1 y columna 2}

\item{o21}{entero: frecuencia observada en la fila 2 y columna 1}

\item{o22}{entero: frecuencia observada en la fila 2 y columna 2}

\item{fcat}{vector de cadenas de texto:  Nombres de fila}

\item{ccat}{vector de cadenas de texto:  Nombres de columna}

\item{estudio}{caracter: Tipo de estudio "T"=transversal, "P"=prospectivo, "R"=retrospectivo}

\item{tablas}{vector de caracter: "E"=frecs. esperadas, "R","S","X", residuos de Pearson, estandarizados y contribucion X2; "F", "C", "T" porcentajes por filas, columnas y totales}

\item{medidas}{valor logico: si es TRUE/FALSE se muestran/omiten las medidas de asociacion}

\item{alfa}{real en (0,1): Nivel de error de los intervalos (alternativa a conf)}

\item{conf}{real en (0,1): Nivel de confianza de los intervalos (alternativa a alfa)}

\item{decs}{entero: Numero de decimales en las salidas}

\item{ptol}{decimal en (0,1): Tolerancia para resumir el valor de p}
}
\value{
Informe analisis de tablas 2x2 mediante tests Chi2 y exacto de Fisher con medidas de asociacion
}
\description{
Analisis de tablas 2x2. Texto intencionadamente sin tildes u otros caracteres especiales por la incompatibilidad de los mapas de caracteres.
}
\examples{
# [1] Formato esperado de la tabla (presencia de enfermedad
# (variable respuesta) vs exposicion a factor de riesgo (var. factor))
#
#             Expuestos      No Expuestos
#--------------------------------------------
# Enfermos        o11           o12
# Sanos           o21           o22
# --------------------------------------------
#
#[1a] Introduccion de frecuencias individuales en estudio transversal,
#     se piden porcentajes por filas
tabla2x2(o11=20, o12=26, o21=60, o22=294, estudio = "T", tablas="F")

#[1b] Los mismos datos introducidos como vector, pidiendo tambien
#     las frecuencias esperadas
tabla2x2(o=c(20, 26, 60,294), estudio = "T",fcat=c("Peso normal","Peso bajo"),
        ccat=c("fuma","No fuma"), tablas=c("F","E","S"))
tabla2x2(frecs=c(20, 26, 60,294), estudio = "T", tablas=c("F","E"))

#[2]  Datos como variables o columnas de un data.frame
w1<-c(1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2)
w2<-c(1,1,1,2,1,1,1,1,2,1,1,1,1,1,2,1,2,2,1,2,2,1,2,2,1,2,2,1,2,2,1,2,1,2,2,2)
tabla2x2(fvar=w1,cvar=w2)
#     Estudio retrospectivo solicitando porcentajes por filas
v1<-c(1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2)
v2<-c(1,1,1,2,1,1,1,1,2,1,1,1,1,1,2,1,2,2,1,2,2,1,2,2,1,2,2,1,2,2,1,2,1,2,2,2)
tabla2x2(fvar=v1,cvar=v2, estudio="R", tablas="F")

}
