% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablarxc.R
\name{tablarxc}
\alias{tablarxc}
\title{Analisis de una tabla de contingencia RxC}
\usage{
tablarxc(
  frecs = NULL,
  fvar = NULL,
  cvar = NULL,
  o = NULL,
  fnum = 0,
  fcat = NULL,
  ccat = NULL,
  tablas = "",
  decs = 3
)
}
\arguments{
\item{frecs}{data.frame: data.frame con frecuencias observadas}

\item{fvar}{vector de enteros o factores: Variable por filas}

\item{cvar}{vector de enteros o factores: Variable por columnas}

\item{o}{vector de enteros: frecuencias de la tabla de contingencia. Requiere especificar el numero de filas con fnum}

\item{fnum}{valor entero: numero de filas cuando se indican las frecuencias a traves del parametro o}

\item{fcat}{vector de cadenas de texto:  Nombres de fila}

\item{ccat}{vector de cadenas de texto:  Nombres de columna}

\item{tablas}{vector de caracter: "E"=frecuencias esperadas, "R","S","X", residuos de Pearson, estandarizados y contribucion X2; "F", "C", "T" porcentajes por filas, columnas y totales}

\item{decs}{entero: Numero de decimales en las salidas}
}
\value{
Informe test para analisis de tablas RxC mediante test Chi2
}
\description{
Analisis de tablas de contingencia RxC (con R y/o C mayores a 2. No valido para tablas 2x2). (Texto intencionadamente sin tildes u otros caracteres especiales por la incompatibilidad de los mapas de caracteres)
}
\examples{
# [1] Introduccion de datos como variables
v1<-c(1,1,2,1,2,1,1,1,1,2,1,1,1,1,1,1,1,1,2,1,1,1,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2)
v2<-c(1,2,3,3,3,2,1,2,1,3,1,2,3,3,3,1,2,1,3,1,2,1,2,3,2,3,1,2,1,1,3,1,1,3,3,1,2,1,1,2,1,2,2,2)
tablarxc(fvar=v1,cvar=v2)
# Obtencion de tablas de frecuencias esperadas y porcentajes por filas
tablarxc(fvar=v1,cvar=v2, tablas=c("E","F"))
# Uso de nombres de categoria
tablarxc(fvar=v1,cvar=v2, tablas=c("E","F"), fcat=c("Trat1","Trat2"),ccat=c("peor","igual","mejor"))

# [2] Introduccion directa de las frecuencias (previamente construir por columnas un dataframe)
peor <-c(8,9,10)
igual<-c(6,8,12)
mejor<-c(4,9,10)
tabla<-data.frame(peor,igual,mejor)
tablarxc(frecs=tabla, tablas=c("E","F","S"), fcat=c("Trat1","Trat2"))

# [3] Introduccion de frecuencias observadas como vector
obs<-c(12,35,13,25,8,10)
tablarxc(o=obs, fnum=2)



}
