% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testf.R
\name{testf}
\alias{testf}
\title{Test de Fisher (F-test) para comparar dos varianzas}
\usage{
testf(s1 = 0, n1 = 0, s2 = 0, n2 = 0)
}
\arguments{
\item{s1}{valor real: desviacion tipica de la muestra 1}

\item{n1}{valor entero: tamano muestra 1}

\item{s2}{valor real: desviacion tipica de la muestra 2}

\item{n2}{valor entero: tamano muestra 2}
}
\value{
Fexp, gl1, gl2 y valor p del test de Fisher para el cociente de varianzas
}
\description{
Permite obtener el intervalo de confianza a partir de una variable o bien de las medidas resumidas (Texto intencionadamente sin tildes u otros caracteres especiales por la incompatibilidad de los mapas de caracteres)
}
\examples{
testf(s1=15, n1=120, s2=12, n2=65)
}
