% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testmcnemar.R
\name{testmcnemar}
\alias{testmcnemar}
\title{test de homogeneidad de dos proporciones apareadas (McNemar)}
\usage{
testmcnemar(
  pre = NULL,
  post = NULL,
  n = NULL,
  n11 = 0,
  n12 = 0,
  n21 = 0,
  n22 = 0,
  fcat = c("+", "-"),
  ccat = c("+", "-"),
  alfa = 0.05,
  conf = 0.95,
  decs = 4,
  delta = NULL,
  beta = 0.2,
  potencia = 1 - beta,
  lbls = NULL
)
}
\arguments{
\item{pre}{vector de observaciones en el pretest}

\item{post}{vector de observaciones en el posttest}

\item{n}{vector de enteros: frecuencias observadas con el formato n=c(n11,n12,n21,n22)}

\item{n11}{entero: frecuencia observada en la fila 1 y columna 1}

\item{n12}{entero: frecuencia observada en la fila 1 y columna 2}

\item{n21}{entero: frecuencia observada en la fila 2 y columna 1}

\item{n22}{entero: frecuencia observada en la fila 2 y columna 2}

\item{fcat}{vector de cadenas de texto:  Nombres de fila}

\item{ccat}{vector de cadenas de texto:  Nombres de columna}

\item{alfa}{real en (0,1): Nivel de error de los intervalos (alternativa a conf) y error de tipo I para la determinacion de n}

\item{conf}{real en (0,1): Nivel de confianza de los intervalos (alternativa a alfa)}

\item{decs}{entero: Numero de decimales en las salidas}

\item{delta}{real: diferencia a detectar}

\item{beta}{real en (0,1): Nivel de error de tipo II para la determinacion de n}

\item{potencia}{real en (0,1): Nivel de potencia deseada (alternativa a beta)}

\item{lbls}{vector de los nombres de variable pre- y post- (llamada desde testp)}
}
\value{
Informe analisis de dos proporciones apareadas mediante test de McNemar, intervalo de conf. para la diferencia y estimacion de tamano muestral
}
\description{
Comparacion de dos proporciones apareadas. Texto intencionadamente sin tildes u otros caracteres especiales por la incompatibilidad de los mapas de caracteres.
}
\examples{
#
#Uso basico de la funcion introduciendo frecuencias observadas conforme a la tabla
#           post+ post-  |
# pre+      n11     n12  |
# pre-      n21     n22  |
#          -----------------
#                        |n
#
# Tabla de frecuencias. Son equivalentes las llamadas
testmcnemar(n=c(27,35,43,20))
testmcnemar(n11=27,n12=35,n21=43,n22=20)

# Determinacion del tamano de muestra para declarar significativa una
# diferencia delta con potencia 1-beta y error alfa
testmcnemar(n11=27,n12=35,n21=43,n22=20,delta=0.05,alfa=0.05,beta=0.15)

# Informacion en forma de vector
xpre<- c(1,1,2,2,1,2,1,2,2,2,1,1,2,2,1,1,2,2,1,1,1,1,1,1,1,NA,2)
xpost<-c(1,2,2,1,2,1,2,2,2,1,1,2,2,1,1,2,2,1,2,NA,1,1,1,2,2,2,1)
testmcnemar(xpre,xpost)

# Determinacion del tamano muestral
testmcnemar(xpre,xpost,delta=0.05, beta=0.15)

}
