% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{testnormal}
\alias{testnormal}
\title{Estudio de la normalidad de una variable}
\usage{
testnormal(
  x,
  obs = TRUE,
  mod = TRUE,
  dens = TRUE,
  ks = 1,
  ky = 1.2,
  qq = FALSE,
  sw = TRUE,
  decs = 3,
  ...
)
}
\arguments{
\item{x}{vector de datos a analizar}

\item{obs}{valor logico: TRUE=representar el histograma}

\item{mod}{valor logico: TRUE=representar el modelo ajustado de densidad normal}

\item{dens}{valor logico: TRUE=Superponer estimador de nucleo de la densidad}

\item{ks}{valor real: factor multiplicador del numero de intervalos calculados (el calculo se hace por la regla de Sturges o la de Freedman-Diaconis)}

\item{ky}{valor real: factor multiplicador de la escala del eje de ordenadas}

\item{qq}{valor logico: representa el diagrama QQ normal}

\item{sw}{valor logico: realiza el test de Shapiro-Wilk}

\item{decs}{valor entero: numero de decimales en la salida}

\item{...}{parametros de configuracion de la funcion grpsggp}
}
\value{
Histograma con densidad normal superpuesta, diagrama probabilistico normal, test de Shapiro-Wilk
}
\description{
Representa el histograma de una variable x, superponiendo la densidad normal ajustada asi como un estimador del nucleo de la densidad. Permite obtener el diagrama QQ y el resultado del test de Shapiro-Wilk
}
\examples{
x=rnorm(500,10,2)
testnormal(x);
testnormal(x,ks=3);
testnormal(x,dens=FALSE)
testnormal(x,mod=FALSE,dens=FALSE,ks=3)
testnormal(x,obs=FALSE)
testnormal(x,col="white")
testnormal(x,sw=TRUE)
testnormal(x,sw=TRUE,qq=TRUE,ks=3)
testnormal(x,sw=TRUE,qq=TRUE,ks=3)
}
