% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testp.R
\name{testp}
\alias{testp}
\title{Test con proporciones}
\usage{
testp(
  x = NULL,
  n = 0,
  x1 = NULL,
  n1 = 0,
  x2 = NULL,
  n2 = 0,
  grupos = NULL,
  x0 = NA,
  p0 = 0.5,
  par = FALSE,
  alfa = 0.05,
  conf = 1 - alfa,
  delta = 0,
  beta = 0.2,
  potencia = 1 - beta,
  decs = 3
)
}
\arguments{
\item{x}{valor entero de casos favorables cuando se trata de una unica muestra o vector binario de observaciones .}

\item{n}{valor entero = tamano de muestra cuando esta es unica o vector con la tabla para test de McNemar}

\item{x1}{valor entero o vector: casos favorables en la muestra 1 o vector de observaciones.}

\item{n1}{valor entero: tamano de la muestra 1.}

\item{x2}{valor entero o vector: casos favorables en la muestra 2 o vector de observaciones}

\item{n2}{valor entero: tamano de la muestra 2.}

\item{grupos}{vector: variable de agrupacion o indicacion a la variable x1 o x2 que haga esta funcion}

\item{x0}{indicador de la categoria a analizar}

\item{p0}{valor de la proporcion para estimar n exclusivamente}

\item{par}{valor logico: indicador de muestras apareadas}

\item{alfa}{real en (0,1): Nivel de error de los intervalos (alternativa a conf)}

\item{conf}{real en (0,1): Nivel de confianza de los intervalos (alternativa a alfa)}

\item{delta}{real en (0,1): Diferencia a detectar (tamano del efecto)}

\item{beta}{real en (0,1): Nivel de error de tipo II (alternativa a indicar la potencia)}

\item{potencia}{real en (0,1): Nivel de potencia (alternativa a indicar el error de tipo II)}

\item{decs}{entero: Numero de decimales en las salidas}
}
\value{
Informe con test e IC, estudio de la fiabilidad de la no significacion y tamano de muestra
}
\description{
Test para una proporcion o estudio de la homogeneidad de dos proporciones binomiales
             independientes o dos proporciones apareadas.
             (Texto intencionadamente sin tildes u otros caracteres especiales por la incompatibilidad de los mapas de caracteres)
}
\examples{

# test con una muestra
testp(x=30,n=70,p0=0.4)
testp(x=30,n=70,p0=0.4,delta=0.05,potencia=0.9)

modos<-c("A","A","A","A","A","A","A","A","A","A","B","B","B","B")
edad <-c(5,4,8,6,7,4,5,4,5,6,9,6,5,4,5,6,4)

testp(x=modos, p0=0.4,delta=0.05,potencia=0.9)
testp(x=edad, x0=6,p0=0.5, delta=0.05,potencia=0.9)
testp(x=as.factor(edad), x0=6,p0=0.5, delta=0.05,potencia=0.9)

testp(p0=0.65,delta=0.05,potencia=0.9)  #solo tamano de muestra
testp(p0=0.5,delta=0.05,alfa=0.05,potencia=0.9)

# 2 muestras independientes ---
fumaH=c("S","N","N","N","N","N","S","N","S","N","S","N","S","N","S","N","N")
fumaM=c("N","N","N","N","N","N","S","S","S","S","S","N","N","S","N","N","S")
testp(x1=fumaH,x2=fumaM)
testp(x1=fumaH,x2=fumaM,x0="S")

sexo=c("H","H","H","M","M","H","M","M","H","H","M","M","M","M","H","M",
       "M","M","M","M","H","M","H","H","H","H")
fuma=c("S","N","S","S","N","N","S","N","S","S","N","N","N","N","S","N",
       "N","N","S","N","S","N","S","N","N","N")
testp(x=fuma,grupos=sexo, x0="S")
testp(x=fuma,grupos=sexo, x0="N")

a<-rbinom(250,1,0.4)
b<-rbinom(250,1,0.6)
testp(x1=a,x2=b,x0=1)



#'# 2 muestras apareadas ----
testp(n=c(12,35,43,20), par=TRUE, delta=0.05)
testp(x1=150,n1=450,x2=34,n2=49, par=TRUE)

fuma1<-c("S","S","S","N","S","N","S","S","S","N","S","N","S","N","S","N","N",
         "N","N","S","N","N","N","S","S","S","S","S","N","N","S","N","N","S")
fuma2<-c("N","N","S","N","N","S","S","N","S","N","S","N","S","N","N","N","N",
         "N","N","N","N","N","S","N","N","N","N","N","N","N","N","N","N","S")
testp(x1=fuma1,x2=fuma2,par=TRUE)

}
