% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testt.R
\name{testt}
\alias{testt}
\title{Test de student con una y dos muestras (independientes o apareadas)}
\usage{
testt(
  m = NULL,
  m1 = NULL,
  m2 = NULL,
  n = 0,
  n1 = 0,
  n2 = 0,
  s = 0,
  s1 = 0,
  s2 = 0,
  par = FALSE,
  m0 = 0,
  grupos = NULL,
  conf = 0.95,
  vac = TRUE,
  alfa = 0.05,
  delta = 0,
  potencia = 0.8,
  beta = 0.2,
  decs = 3,
  grf = TRUE
)
}
\arguments{
\item{m}{vector o valor real: vector de datos cuando se indica una muestra o la variable a analizar en el caso de muestras independientes. Media de la variable a analizar, previamente calculada, cuando se indica una sola muestra con la informacion resumida}

\item{m1}{vector o valor real: vector de datos de la primera muestra cuando se indican dos muestras apareadas. El contraste sera Media(m1)-Media(m2)=m0, o medias muestrales cuando se indican dos muestras independientes con las medidas de sintesis}

\item{m2}{vector o valor real: vector de datos de la segunda muestra cuando se indican dos muestras apareadas. El contraste sera Media(m1)-Media(m2)=m0, o medias muestrales cuando se indican dos muestras independientes con las medidas de sintesis}

\item{n}{valor entero: tamano muestral cuando se indican los datos resumidos de una sola muestra}

\item{n1}{valor entero: tamano de la muestra 1 cuando se indican los datos resumidos de dos muestras independientes,}

\item{n2}{valor entero: tamano de la muestra 2 cuando se indican los datos resumidos de dos muestras independientes}

\item{s}{valor real: desviacion tipica de la variable en el test con una muestra.}

\item{s1}{valor real: desviacion tipica de las muestras 1 para la comparacion de 2 muestras independientes especificadas por sus medidas de sintesis.}

\item{s2}{valor real: desviacion tipica de las muestras 2 para la comparacion de 2 muestras independientes especificadas por sus medidas de sintesis.}

\item{par}{valor logico: si los tamanos de m1 y m2 son iguales se asumen muestras apareadas, pero si par=FALSE se asumen independientes}

\item{m0}{valor real: valor a contrastar en el test de una muestra o magnitud de la diferencia (efecto bruto) en los test con dos muestras}

\item{grupos}{variable binaria o factor con dos niveles: variable de agrupacion en la comparacion de dos muestras independientes especificadas con los datos individuales de cada caso}

\item{conf}{valor real < 1: nivel de confianza para la elaboracion del IC para la estimacion de la media o del tamaño del efecto}

\item{vac}{valor logico: TRUE=se trata de una variable aleatoria continua; FALSE= la variable es discreta y se aplica cpc_ Por defecto = TRUE.}

\item{alfa}{valor real < 1: error alfa (parametro alternativamente al nivel de confianza, en tanto por uno). Por defecto =.05.}

\item{delta}{valor real: tamano del efecto a detectar, es decir, a declarar significativo con la potencia deseada}

\item{potencia}{valor real <1: potencia deseada para estudiar la fiabilidad de la decision por la hipotesis nula y el tamaño de muestra.}

\item{beta}{valor real: error de tipo II, parametro alternativo a la potencia}

\item{decs}{valor entero: precision decimal para la salida de resultados. Por defecto = 4.}

\item{grf}{valor logico: Si TRUE/FALSE se genera/omite la salida grafica}
}
\value{
Informe con medidas descriptivas, test de normalidad (si se aportan datos individuales), test del cociente de varianzas de Fisher (si procede), t-test con estimacion del tamano del efecto bruto, estudio de la potencia y estimacion de tamano muestral.
}
\description{
t-test con una y dos muestras, independientes o relacionadas  (se han omitido acentos deliberadamente por la incompatibilidad de caracteres de texto)
}
\examples{
# [A] Test con una muestra
# [A.1] Con los datos individuales
datos<-c(76,54,12,47,13,15,25,14,19,32,7)
testt(m=datos, m0=25)
testt(m=datos, m0=29,delta=3)
testt(m=datos, m0=29,delta=3,vac=FALSE)

# [A.2] Con informacion muestral sintetizada
testt(m=37, s=5, n=158, m0=25)
testt(m=37, s=7, n=158, m0=36,delta=3,potencia=0.95)

# [B] Test con dos muestras independientes
# [B.1] 2 muestras independientes con datos individuales
sexo<-c( 1,2,2,2,1,1,2,2,1,1)
peso<-c(54,64,76,84,45,74,76,95,63,62)
testt(m=peso, grupos=sexo)
fuma<-c("si","no","si","no","no","no","no","si","no","si")
testt(m=peso, grupos=fuma,delta=5,potencia=0.95)

grupo1<-c(12.5,7.4,8.3,4.6,5.1,7.8,9.2,4.6)
grupo2<-c(8.7,14.8,13.5,16.1,7.1,19.2,21.5,22.4,18.7)
testt(m1=grupo1, m2=grupo2)

# [B.2] 2 muestras independientes con datos sintetizados
testt(n1=123, m1=25, s1=6, n2=87, m2=20, s2=8)

# [C] Test con dos muestras relacionadas
# [C.1] 2 muestras apareadas con datos individuales
pre_tratamiento<-c(3.2,4.5,1.7,2.6,1.7,4.3,2.1,3.8,4.9,5.1,NA)
post_tratamiento<-c(3.8,4.1,2.2,3.1,2.7,9.3,7.9,3.1,5.7,5.3,NA)
testt(m1=pre_tratamiento, m2=post_tratamiento, par=TRUE)
testt(m1=pre_tratamiento, m2=post_tratamiento, par=TRUE, delta=2,m0=0.1)
testt(m1=pre_tratamiento, m2=post_tratamiento, par=TRUE, delta=2,m0=0.1, vac=FALSE)

# [C.2] 2 muestras apareadas con datos sintetizados (es el test con una sola
#       muestra, habitualmente m0=0)
testt(m=0.65, s=1.2, n=17)

# [D] Estudio de la fiabilidad por H0 y estimacion del tamano de muestra (basta
#    anadir el parametro delta a cualquiera de las opciones anteriores)
testt(m=peso, grupos=fuma, delta=5, potencia=0.95)
testt(m1=pre_tratamiento, m2=post_tratamiento, delta=0.5, potencia=0.85)
testt(m1=pre_tratamiento, m2=post_tratamiento, delta=0.5, beta=0.15)

}
