% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testwx.R
\name{testwx}
\alias{testwx}
\title{Test no parametricos con dos muestras independientes (Wilcoxon/Mann-Whitney) y apareadas (Wilcoxon)}
\usage{
testwx(
  m1 = NULL,
  m2 = NULL,
  par = FALSE,
  m = NULL,
  grupos = NA,
  grf = TRUE,
  alfa = 0.05,
  conf = 1 - alfa,
  decs = 3,
  mess = -1
)
}
\arguments{
\item{m1}{vector: vector de datos de la primera muestra cuando se indican dos muestras apareadas y tambien valido para muestras independientes.}

\item{m2}{vector: vector de datos de la segunda muestra cuando se indican dos muestras apareadas y tambien valido para muestras independientes.}

\item{par}{valor logico: si los tamanos de m1 y m2 son iguales se asumen muestras apareadas, pero si par=FALSE se asumen independientes}

\item{m}{vector: vector de datos a contrastar en formato longitudinal. Es preciso especificar el vector grupos para segmentar a este vector}

\item{grupos}{vector factor con dos niveles: variable de agrupacion en la comparacion de dos muestras independientes con valores dados en m}

\item{grf}{valor logico: si grf=FALSE se omite la salida grafica}

\item{alfa}{valor real < 1: error alfa (parametro alternativamente al nivel de confianza, en tanto por uno). Por defecto =.05.}

\item{conf}{valor real < 1: nivel de confianza para la elaboracion del IC para la estimacion del efecto}

\item{decs}{valor entero: precision decimal para la salida de resultados. Por defecto = 4.}

\item{mess}{valor entero: -1 desactiva mensajes de aviso (ver documentacion de Rs para options(warn = valor))}
}
\value{
Informe con estadisticos de orden, rangos, resultado del test y estimacion del tamano del efecto
}
\description{
Test de homogeneidad no parametricos para dos muestras independientes (test de Wilcoxont/Mann-Whitney con aprox. a la normal) y con dos muestras apareadas (Wilcoxon). Se dan algunas medidas de tamano del efecto. (Omision deliberada de tildes por compatibilidad)
}
\examples{
#[A] Muestras independientes
#[A.1] Como vectores independientes (distinto tamano)
y1<-c(78,64,75,45,82,55,48)
y2<-c(110,70,53,51,63,87)
testwx(m1=y1,m2=y2)

#[A.2] Como vectores independientes (con el mismo tamano)
y1<-c(78,64,75,45,82,55,48)
y2<-c(110,70,53,51,63,87,99)
testwx(m1=y1,m2=y2,par=FALSE)

#[A.3] En formato longitudinal (vector de valores y vector de agrupacion)
y<-c(78,64,75,45,82,55,48,  110,70,53,51,63,87)
g<-c(1,1,1,1,1,1,1,2,2,2,2,2,2)
testwx(m=y,grupos=g)

#[B] Muestras apareadas (y1 e y2 deben tener el mismo tamano)
y1<-c(78,64,75,45,82,55,48)
y2<-c(110,70,53,51,63,87,99)
testwx(m1=y1,m2=y2)

}
