% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Europe}
\alias{Europe}
\title{Durées de travail en Europe}
\format{
A data frame with 25 observations on the following 2 variables.
\describe{ \item{Pays}{a factor with the some of the European
countries as levels} \item{Duree}{weekly cumulative work duration} }
}
\description{
This dataset provide mean weekly cumulated work durations for several
European countries.
}
\details{
The duration is given in hours
}
\examples{

data(Europe)

}
\references{
F. Bertrand, M. Maumy-Bertrand, Initiation à la Statistique avec
R, Dunod, 4ème édition, 2023.
}
\keyword{datasets}
