\name{Europe}
\alias{Europe}
\docType{data}
\title{Durées de travail en Europe}
\description{
This dataset provide mean weekly cumulated work durations for several European countries.
}
\usage{Europe}
\format{
  A data frame with 25 observations on the following 2 variables.
  \describe{
    \item{\code{Pays}}{a factor with the some of the European countries as levels}
    \item{\code{Duree}}{weekly cumulative work duration}
  }
}
\details{The duration is given in hours}
\source{F. Bertrand, M. Maumy-Bertrand, Initiation à la Statistique avec R, Dunod, 2010.}
\examples{
data(Europe)
}
\keyword{datasets}
