% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.hist.R
\name{panel.hist}
\alias{panel.hist}
\title{Histrogammes}
\usage{
panel.hist(x, ...)
}
\arguments{
\item{x}{Un vecteur numérique}

\item{\dots}{Des arguments à transmettre à la fonction qui créé les
histogrammes}
}
\description{
Sert à représenter des histogrammes dans les graphiques matriciels
}
\details{
Cette fonction s'utilise avec la fonctions graphique pairs.
}
\examples{

data(Mesures5)
pairs(Mesures5,diag.panel="panel.hist")

}
\references{
F. Bertrand, M. Maumy-Bertrand, Initiation à la Statistique avec
R, Dunod, 4ème édition, 2023.
}
\seealso{
\code{\link{pairs}}, \code{\link{hist}}
}
\author{
Frédéric Bertrand\cr \email{frederic.bertrand@lecnam.net}\cr
\url{https://fbertran.github.io/homepage/}\cr 
Maumy-Bertrand\cr \email{myriam.maumy@ehesp.fr}\cr
\url{https://www.ehesp.fr/annuaire/enseignement-recherche/myriam-maumy/}
}
\keyword{univar}
