% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.hist.R
\name{panel.hist}
\alias{panel.hist}
\title{Histrogammes}
\usage{
panel.hist(x, ...)
}
\arguments{
\item{x}{Un vecteur numérique}

\item{\dots}{Des arguments à transmettre à la fonction qui créé les
histogrammes}
}
\description{
Sert à représenter des histogrammes dans les graphiques matriciels
}
\details{
Cette fonction s'utilise avec la fonctions graphique pairs.
}
\examples{

data(Mesures5)
pairs(Mesures5,diag.panel="panel.hist")

}
\references{
F. Bertrand, M. Maumy-Bertrand, Initiation à la Statistique avec
R, Dunod, 3ème edition, 2018.
}
\seealso{
\code{\link{pairs}}, \code{\link{hist}}
}
\author{
Frédéric Bertrand\cr \email{frederic.bertrand@math.unistra.fr}\cr
\url{http://www-irma.u-strasbg.fr/~fbertran/}\cr 
Maumy-Bertrand\cr \email{myriam.maumy@math.unistra.fr}\cr
\url{http://www-irma.u-strasbg.fr/~mmaumy/}
}
\keyword{univar}
