\name{Quetelet}
\alias{Quetelet}
\docType{data}
\title{Indices de Quetelet}
\description{
This dataset provide heights and masses for computing biomass indices.
}
\usage{Quetelet}
\format{
  A data frame with 66 observations on the following 3 variables.
  \describe{
    \item{\code{sexe}}{a factor with the sex of the individual}
    \item{\code{poids}}{weight of the individual}
    \item{\code{taille}}{height of the individual}
  }
}
\details{The weight is given in kg and the height in cm}
\source{F. Bertrand, M. Maumy-Bertrand, Initiation à la Statistique avec R, Dunod, 2010.}
\examples{
data(Quetelet)
}
\keyword{datasets}
